requirements-dev.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
djangorestframework==3.12.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
iniconfig==1.1.1
oauth2client==4.1.3
packaging==20.9
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
